+++ lib/lingua.rb
@@ -2,6 +2,7 @@
 
 require_relative 'lingua/version'
 require_relative 'lingua/lingua'
+require_relative 'lingua/language'
 
 module Lingua
   class Error < StandardError; end