+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.0
+
+- Add language predicate methods (`lang.french?`, `lang.english?`, etc.)
+- Add `iso_code` alias for `to_iso6391`
+
 ## 0.3.0
 
 - Add `Lingua::Language[]` lookup by name, ISO 639-1 or ISO 639-3 code